Moved AccoutSettingsRemote + tests + basic models to WPKit framework
