+++ WordPress/Classes/Services/AccountSettingsService.swift
@@ -1,5 +1,6 @@
 import Foundation
 import Reachability
+import WordPressKit
 
 let AccountSettingsServiceChangeSaveFailedNotification = "AccountSettingsServiceChangeSaveFailed"
